+++ src/init.server.lua
@@ -13,7 +13,7 @@ local Util = require(IPS2DevKit.Util)
 local App = require(script.App)
 
 local NAME = "IPS2DevKit"
-local VERSION = "v0.7.0"
+local VERSION = "v0.8.0"
 
 local toolbar = plugin:CreateToolbar(NAME)
 local widget = Util.CreateWidget(plugin, NAME)